+++ RestCountry/ui-countryList/src/main/java/com/iliadmastery/ui_countrylist/components/CountryListItem.kt
@@ -34,7 +34,7 @@ import com.iliadmastery.ui_countrylist.R
 @Composable
 fun CountryListItem(
     country: Country,
-    onSelectCountry: (Long) -> Unit,
+    onSelectCountry: (Int) -> Unit,
     imageLoader: ImageLoader,
 ) {
     Surface(
